feat: add cloud sync hook and panel component
